// src/api/stetsom/model/templateAttrInput.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export interface TemplateAttrInput {
  attribute_id: string;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
}